requirements.txt
# Flask and web server dependencies
Flask==2.3.2
python-dotenv==1.0.0
Werkzeug==2.3.6
jinja2==3.1.2
SQLAlchemy==2.0.19

# Data processing and scientific libraries
opencv-python==4.8.0.74
pandas==1.5.3
numpy==1.26.3
tensorflow==2.16.1
h5py==3.13.0
Pillow==10.0.0
scikit-image==0.21.0
matplotlib==3.8.2
imutils==0.5.4

# API and HTTP libraries
requests==2.31.0
gdown==5.2.0
roboflow==1.1.0
inference-sdk==0.36.1

# AI/ML libraries
google-generativeai==0.3.1
langchain-core==0.1.5
supervision==0.18.0

# NLP and text processing
nltk==3.8.1
spacy==3.7.2
emoji==2.8.0
python-dateutil==2.8.2
dateparser==1.1.8
textblob==0.17.1
langdetect==1.0.9
markdown==3.4.3
deep-translator==1.11.4

# Text-to-speech and speech recognition
gtts==2.3.2
SpeechRecognition==3.10.0

# Additional utilities
# Note: sqlite3 is part of Python's standard library, no need to pip install it
# Note: uuid is part of Python's standard library, no need to pip install it

# Spacy language model - install with: python -m spacy download en_core_web_sm 